examples/BraTS2023/mocked_metrics/project/requirements.txt
PyYAML~=5.3
typer
SimpleITK>=2.1.0
numpy~=1.24.2
